Update: Register form

+++ laravel-coffee-website/database/migrations/2014_10_12_000000_create_users_table.php
@@ -21,7 +21,7 @@ return new class extends Migration
             $table->string('user_foto')->nullable();
             $table->tinyInteger('user_status')->nullable();
             $table->text('alamat')->nullable();
-            $table->integer('no_hp')->nullable();
+            $table->bigInteger('no_hp')->nullable();
             $table->string('email')->unique();
             $table->timestamp('email_verified_at')->nullable();
             // $table->tinyInteger('level');
